+++ src/sheets/core.ts
@@ -200,4 +200,128 @@ sheet.glyphs.set("tower", (canvas: SVGContainer) => {
     return group;
 });
 
+sheet.glyphs.set("wyke-1", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-1");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-2", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-2");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-3", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-3");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-4", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-4");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.use(rectsym).dmove(50,50);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-5", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-5");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.use(rectsym).dmove(50,50);
+    group.use(rectsym).dmove(6.25, 6.25).scale(0.75, 6.25, 6.25);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-6", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-6");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.use(rectsym).dmove(50,50);
+    group.use(rectsym).dmove(6.25, 6.25).scale(0.75, 6.25, 6.25);
+    group.use(rectsym).dmove(56.25, 6.25).scale(0.75, 56.25, 6.25);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-7", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-7");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.use(rectsym).dmove(50,50);
+    group.use(rectsym).dmove(6.25, 6.25).scale(0.75, 6.25, 6.25);
+    group.use(rectsym).dmove(56.25, 6.25).scale(0.75, 56.25, 6.25);
+    group.use(rectsym).dmove(6.25, 56.25).scale(0.75, 6.25, 56.25);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
+sheet.glyphs.set("wyke-8", (canvas: SVGContainer) => {
+    const group = canvas.symbol().id("wyke-8");
+    const rectsym = group.symbol().id("rect")
+        .rect(50, 50)
+        .fill("none")
+        .attr("data-playerfill", true)
+        .stroke({width: 1, color: "black"})
+    group.use(rectsym);
+    group.use(rectsym).dx(50);
+    group.use(rectsym).dy(50);
+    group.use(rectsym).dmove(50,50);
+    group.use(rectsym).dmove(6.25, 6.25).scale(0.75, 6.25, 6.25);
+    group.use(rectsym).dmove(56.25, 6.25).scale(0.75, 56.25, 6.25);
+    group.use(rectsym).dmove(6.25, 56.25).scale(0.75, 6.25, 56.25);
+    group.use(rectsym).dmove(56.25, 56.25).scale(0.75, 56.25, 56.25);
+    group.viewbox(-1, -1, 102, 102);
+    return group;
+});
+
 export { sheet as CoreSheet };